[WIP] Make the Bluespace Harpoon Require Charges/Battery to Use!
TL;DR BS Harpoon is really OP with the literally infinite uses as a result of it being able to be used without a power source required.

This finally wrestles control of this eldritch thing and forces it to use power like any other normal energy-using piece of technology.

IMO, BS Harpoons are still really broken and shouldn't be given to RD's right off roundstart, but if I can get this to work, it'll fix most of the issues I have with it being literally free uses.