Refactors trash pile prob chances
Chances are now scaled in relation to the highest. Might improve performance a bunch considering the old chances essentially had a minimum reroll failure rate of 94%.

From what I've gathered, the way this probability picking works is that it picks a random item off the list, and then uses the probability to determine if the roll was successful to receive the item, and with a 5% probability on the highest tier, this would mean a 95% failure rate for each reroll.

Also reduces emag probability to not be on par with glowsticks and pizza vouchers.