Revert strafing speed nerf. (#7618)
A 1.5x modifier was added to strafing speed without much of any info. This is not something that should had been added without at least clearly saying it in the PR that introduced it.

Maybe it could comeback for shitty actuators, but seeing as we only have a single actuator variant at the moment, this has to go.