Phroronguarded Work Attire
If it's thick enough to protect from cold, thorns, and other things, it probably should not absorb phoron. All gloves, boots, and clothings that are thick enough to protect against these things should be resistant to phoron contamination. Items with a permeability_coefficient of 0.05 or less now have PHORONGUARD.

Unsure if I should be using flags = PHORONGUARD or item_flags = PHORONGUARD, so setting this as a draft. Will look over it more later.