Telecoms fixes.
Fixed a mismarked area on S1

Moved an intercom in Telecoms, and made sure the airlocked worked BOTH ways.

Changed the density of the subspace broadcasters from 1 to 0. It's a flat disk that lays flush with the ground, there is no reason that one could not reach over or step on it if need be. In Virgo's case, this allows access to the Telecoms hub without having to deconstruct anything; If the Chief Engineer wants to set up a relay offworld or on a ship, he need access to the hub.