Add Surgery and Resleeving Rooms to Robotics!
Adds two surgery rooms and a resleeving room to robotics to make things a little more private!

If people still want to be treated like nothing but machinery in the public view, though, it's not difficult to setup an optable in public still - but this allows people to privately scene or conduct longer, more detailed synth RP + resleeves without clogging up chat/blocking Robotics, and especially because Robotics used to be so visible from literally everywhere.

General overview screenshot here:
![](https://i.imgur.com/hVHneG3.png)