Talon + SME Edits & QOL
Added a NIFshop and NIFsoft boxes to the Talon.

Added a preset console to the SME engine to allow borgs and the less financially well-off crew to access the supermatter monitor for more advanced engine setups without having to run back and forth between the CE's computer and the engine monitoring room.

Switched the Talon's engine to something more fitting for a spaceship and with a much easier to obtain fuel source. It even has a couple Radioisotope Thermoelectric Generators to power the ship through those long space voyages where the crew has to be in cryo! Updated signage to warn of the new radiation hazard.

Added a Cyborg Recharging station to the Talon Generator Core room. A last bastion for any synthetics in a zero-power situation and the top priority for the RTG's to recharge before their power gets to the SMES units.