DNM: Added New item (Cyber Tech PCM)
Fixed a typo in the Chameleon.dm

Added the Cyber Tech PCM, an invisible gas mask. This sensor suite conforms to the user and creates an invisible nano-weave around vital facial components without obstructing their sense of fashion or ability to eat. Comes at a cost.

TODO:
Make this item only work for FBP's and have an EMAG interaction. I'm new to all of this so that may take awhile. Posting for visibility and so I can ask for help on my TODO's.