Update mob spawners/spawns
Teppi can rarely spawn on Virgo 3b now after an incident where they were released and able to survive. 

This also makes it so that gaslamps, geese, and teppi are all in the same spawn pools and such, with gaslamps being 3 times as likely as geese, and teppi being very rare, this will result in the odd goose outside of the wilderness, but, geese are easy to beat up and slow, so, that's probably fine, especially considering how few spawns there are outside of the wilderness.