The final updates
Adds a new random spawner for coins with a fail chance (so the coins aren't always in the same place)

Addsome missing vending machines to the random vending machine spawner 

Adds a sound to starting and stopping in ships (that only plays at most once every 10 seconds.)

And sets it up so that planets and ships can have a sound when visited.

A bunch of little map tweaks to polish things up!