Sensible Dragon Tweaks for Virgo
A few minor tweaks to make the Dragon a more reasonable boss for Virgo purposes. Drastically cutting it's base speed and lowering health by 25%. However also fixing it's issue where it chokes to death so it can't just be cheesed. Still a bandaid fix, it's more suited for V2, frankly.