Revamps adminbus portals
Automates the connection process a little based on how bluespace rifts work.

They also allow the portals to be linked to other portals without recursive portal-ing, or plopping people into walls or other obstructions. (It doesn't account for other conditions though, so if you put a portal next to lava or a horrible sky fall, that's on you). It only puts them immediately adjacent to linked portals or otherwise obstructed target locations.

Basically though, if there is an unlinked portal,  a staff member can click it, they get a dialogue box with some instructions. They can then fly off to where they want the linked portal to go, click the appropriate options, and it will automatically set it up for you. You can select a sister portal, or just a target. Selecting a portal in this way also brings up options for the portal's sprite too, in case you don't have your own~

This should make setting up events that use the portals a little easier.