Change painting persist decay
Instead of paintings lasting 1000 shifts before being deleted

Now the system will store up to 1000 paintings, and remove the one at the top of the list (the earliest/oldest one) first when the maximum number of paintings is exceeded. 

Additionally, examining a painting that has been stored will move it to the bottom of the list, effectively renewing its lease in the system. 

This should make it so that paintings aren't just arbitrarily being removed for being old, and, the paintings that actually get looked at should stick around.