First version of map compilation
Currently, the script will look inside of data/map.txt. The server itself can set the next map in rotation in here, either by admin override, vote or automatic rotation.

The currently available maps that are available, are set in config/map_rotation.txt. This would allow, to completely disable a specific map from the rotation, for example for event reasons.

The script will do the rest once called.
Note: The first time running the script, might prompt to accept that it will run in trusted mode. It is very recommended to also check "Remember this decision" in order for this to work.