Adjust voice/emote sound settings
Makes it so that the voice and emote sounds do not use the ignore_walls variable, and adjusts the extrarange variable to reach normal screen view range.

This was a frustrating decision to make. I used the ignore_walls variable originally, because I wanted to make sure you were only hearing people you could see.

But. The way it works is, that first, it finds the turf of the sound source, then it checks to see who that turf can see, and delivers the sound to those ones.

This means that it will never deliver those sounds to anyone inside of micro holders, lockers, tummies, or anyone inside of anything, really. You can't even hear yourself, because the turf whatever you are in is on can't see you!

That's silly, but that's how it works. Maybe someone can make it better someday. 

SO, you will be able to hear say, whisper, and me sounds through walls now. I did ensure that that these sounds only go the standard view distance though, so you shouldn't hear people who would be off screen from you. (Whispers can only be heard from 2 tiles away)

I figured that, being as a major point of the sounds is knowing when you got a message, so not hearing them in very common situations is more of a priority than hearing some for people who are near you but out of sight. I'd prefer to have my cake and eat it too, but I don't know how to do that in this case, so! It be how it do.