Tunnel Snakes Rule
Makes it so that certain squiggly mobs can go into mouse holes even though they fall outside what's usually allowed!

These mobs also get the option to eat anything else that might be in the tunnel with them!

The special mobs include:
catslugs
morphs
protean blobs
slimes (which promethean slimes are a subtype of)

The thinking is these are squishy things that could reasonably be made to fit, or otherwise fit thematically!

I would like to make it so snakes could do this too, but we should get some vore enabled snakes first!