Implements handle_adrenaline()
When a human mob feels like they're going to be in trouble soon, they can prepare for trouble with "Toggle adrenaline response"

This readies them to inject epinephrine in their bloodstream after getting sufficiently hurt, which will allow them to run away as if on hyperzine, stand up immediately from knockdown and ignore pain.

This is not without cost though, for this can only trigger once/5 minutes, and timer starts when the last of metanephrine clears your system. Too much adrenaline will leave you shaking and unable to stand.