Makes the mining drill use lists. Updates some ore names.
Uses the same magical anti-lag lists as the other mining items!
Some mining drill items had incorrect names, such as glass being 'silicates' which isn't an ore, which is essential for this change. This changes that.

This is purely a name change and users won't see anything different.

Also removed rutile as it was duplicated in the list. Presumably due to Polaris adding rutile and it never being removed.