Gives removed_organ a name. Fixes robo surgery.
Allows you to properly perform surgery on FBP without issues.

Having a tgui input be in can_use results in really weird stuff happening. Even before this PR, it'd give you a input menu, you'd cancel it, and it'd then go into the hemostat step when you obviously didn't want to do that.

This fixes that.