Healing mobs aren't racist against all other simplemobs anymore
Adds the 'nom_mob' verb to simplemobs.

The AI that handles the healing variant of the big dragon and the leopardmander will ignore mobs with this variable. 

Additionally, made it so that this AI will ignore simplemobs with no AI, or an AI that is not hostile.

also makes it so they will ignore simplemobs who are player controlled even if their mob is traditionally hostile