Updates Traveller's Guide with new species lore
Replaces blurbs with cataloguer information when able.

Otherwise, using wikipedia blurbs.

Added Rapala and Altevians and Drones and Cyborgs.

Edited positronic lifespan to follow Synthetic lore on wiki

Wrote custom text for drones and cyborgs, in spirit of the Synthetic Page.

Deleted ancient "TODO" commented out lore as that will never be relevant again. We'll just use the wiki where needed.

Government, locations etc untouched for now. They're on the TODO list.

Also TODO: potentially edit language so it is consistent across blurbs.

Also added an OOC note on the main page to warn that this book contains IC bias, and to consult the wikipedia's backstory entry for objective lore.