Mob tweaks and fixes
The mob cooldown thing from upstream made everything with a positive number in the variable be ultra slow. I'm not sure how it was that the variable was broken previously considering that the higher the number was the slower things went, but now it's ultra slow or nothing! cringe.

Adjusts some of the mobs I've made that are supposed to not be slow to still not be slow, as well as touches up a couple of things with squirrels