One-way conveyor switches
Makes it possible to make one-way switches that are not limited to being mapper only, inspired by other codebases that do similar things (TG and Yogs I know do this.)

You can make a switch into a one-way (or make a one-way back into a two-way) by wrenching it with it's service panel open.

The original one-way lever is changed to use the new system as well, so maps dont need to remove the levers or anything. Aside one which for some reason was a var-edited one-way switch to be reverse only?? I commented out the var-edit so it wouldn't error.