Tweak(mapping): Significantly reworks sci-aerostat solars airlocks
-Fixes south-east sci-aerostat solar airlock starting Unbolted

While fixing this bug, I remembered that the aerostat commonly suffers from the fact that late-join attempts to power it up fail from bolted airlocks. This was fine, due to there being an unbolted one in the south-east.

However, said unbolted south-east one didn't even work at all to cycle in/out, this was due to absence of airlock door labels' absence.

However, adding those labels made the solars tracking computer inaccessible in event of aerostat running out of power.

Therefore, alternate solution meant moving the tracking computers out of the airlock, which led to a significant rework of the 4 aerostat airlocks.