fix(survivorpod): Fixes ghost query preferences
Fixes unreported bug where survivor pods would not send a query.
This was caused due to imperfect fix to inability to opt out from it for the round.

tweak(survivorpod): Makes it clear when the pod was clicked on

Does this by making a visible_message when clicked and changing its description