feature(TGUI input): Adds preference toggle for input lock
Adds a new verb under the Preferences verb tab, "Togle-TGUI-Input-Lock" that when pressed/called from command line invert the present preference from TRUE to FALSE and vice versa, saves it to persist and notifies the user that they have enabled/disabled it and what this setting does.

This was done using a simple var = !var rather than the toggle_preference() proc as none of the TGUI preferences have a preference datum. I believe this is due to the way they're positioned in the character set-up window (they're on the top rather than the bottom). As such, I felt it best not to create a datum just for this.

The reason we need this is that although inserting newlines is great when RPing in a slower paced, heavier setting - fast-paced interactions become unnecessarily clunky with the input lock and opening character setup to toggle it back and forth is tiresome. This way, players can just press a button during gameplay or type "Toggle-TGU" and press space to autocomplete.