hotsprings
Adds a new redgate map: Hotsprings! It's a large snowy map, most of which is a fairly empty snowfield. However, there are a few buildings dotted around, and in the top corner, there are some cosy hotsprings to enjoy. There are few fun hidden things to find too, along with some toys. Be warned, the place is pretty cold.

Also adds warm rocks! These are just simple rocks that do not allow atmos to pass through, like a atmos retention field. They're basically just for mapping to separate warm areas from cold ones without atmos messing it up, and they're used here to make the hotsprings work.

Also, just a warning, this removes the commenting out of the other redgate maps from the stellar delight defines file, bringing them all back into rotation. I assume this will be happening when this map is merged, so I thought it'd be easier to just do it here.