Rain City
Adds a new redgate map: Rain City. This map is a single Z-level, but very very dense, with close to 70 buildinds, most of which are unique (about 18 of them are homes, the rest are all different). It has a LOT of randomly generated loot. Alley ways have random vore mobs, whilst buildings and streets have a variety of preset ones. Makes use of the new fake_sun weather feature to have it raining outside.

In addition to this, it also adds a whole bunch of neon signs, and coloured bulbs to go with them.

It also adds a new darker pavement type called City Slab, a new road turf and dashed line decals to mark the road.