Wolftaur mobs
Added 10 new wolftaur mobs, including 5 different colours with a nude and clothing version of each. They are retaliate mobs that are hard to escape from once they eat you and digest quickly.
They also spawn in a random size from 1 to 1.4, even if it makes the sprites a bit wonky sometimes. The default /wolftaur mob spawns one of 9 of the icon states, but each can also be spawned individually without randomisation.

Also adds unique factions to catgirls, lamias and cookiegirls, so that if they are spawned by the retaliate spawner, they can be provoked into eating you. This is I believe the intended behaviour with the retaliate spawner, I can remove it if not. They will still be safe if they are in their default passive forms.