Redgate train map
Adds a new redgate map: A moving train. A long, linear map with two levels, all aboard a moving train. Has a few things to find and mobs to eat/get eaten by. lots of different rooms, mostly to vibe in. Very very deadly to fall off of, but has railings everywhere except one roof section.

Adds a new death step_trigger, which deletes the body of the person who steps on it and gives them a message about how they died. Also logs and messages the admins so that it can be tracked easily.

Adds a new subtype of ticket printer, which prints permits instead, only little pink tickets. Idea is for them to be used for train tickets if someone wanted to RP a train conductor, but made them broad so that they can be used for anything. Does not add the message to the end of round report, but does notify the staff when a ticket is produced to prevent abuse.