Bluespace Collars
Added two new bluespace collars. They both control the size of the wearer via signalers. The first one allows you to select the size by clicking on it, and is locked to a code and frequency. The second is made by modifying the first, by adding a signaler to it. This one does not allow you to set the size or code on it, and is only locked to frequency. The size you are resized to is determined by the code that is sent from a signaler.
Adds these to various lists, such as trader items, RnD and maint loot.

Also added a new "random bluespace item" random item spawner for mapping use. It is currently not mapped in anywhere.