Vore Panel Health Check
Added a new option to humanoid mobs in vore belly contents: Health Check.

Using this on a mob in your belly returns a very simple readout of their health as a percentage. It also gives notice when they are weakened, paralysed, blinded or sleeping, so that the pred can know whether or not they'll be able to emote/hear emotes.