Ports cooker changes
Ports cooker changes from RogueStar by VerySoft.

Added a "Toggle Safety" option to cooking appliances. This is off by default, can be toggled with a verb on the appliance. When on, instead of food burning when it reaches that point, it is instead ejected from the machine safe to eat.

Significantly sped up the speed that machines reach cooking temperature at. Generally reducing the wait time from about 10 minutes to 2 minutes.

Reduced cooking time of most food significantly.

Main goal of this to make it less frustrating to produce large spreads of food.