Maint Pred Availability
Added a new maint pred ghostpod for redgates. I have added two of these to redgates that I made only (for now). They are normal ghost pods but are always available from the start of the shift. The can be found in the find ghost pod verb and are called creature holes.

Made a lot more mobs in my own redgate maps ghost joinable. Generally one of each type of mod that isn't a humanoid.

Added a 25% chance for each mob to become ghost joinable from the semi-random vore mob spawners used in redgates. This applies to the passive and retaliate spawners only.

Also made the other redgate maps available on the SD again.