Reduces vore death timer on autoresleever (#15948)
Changed the timer on the autoresleever from 15 minutes to 5 minutes, a similar amount of time to being able to make a transcore announcement.

The idea behind this reduction is that the autoresleever is already considerably more popular than the resleever, and it does not make sense to inconvenience people for both engaging in their kinks and not wanting to get medical involved in their death. The normal resleever remains the best option for non-vore deaths, and for people who just want to offer that kind of interaction to medical.