Temporarily disables sleeping on simple mobs
Due to a bug that causes them to never wake up again, the previous attempt at fixing this caused death to stop working, so I'll leave it like this until a solution is found whilst leaving the option in for carbon mobs.

Not the most elegant solution but hopefully only temporary, seeing as simple mobs may be even more accessible to players soon.