Flippable beds and sheets
Added a verb to beds, bedsheets and sub-objects to allow them to be flipped to face the opposite direction (if they have directional sprites available). This applies to chairs too, because why not. If it's possible for me to just not display the verb for non-flippable subtypes, let me know how!

Also swaps out some beds on the Stellar Delight to face the other direction so that their pillows are against the wall.