Winter event mobs
Added christmas catslug variant.

Added a grey scaled drake variant of the dragon, specifically intended to be played by the creator of the art for the upcoming event (and future events), as it's not quite polished yet for every day use. Though they are fine with it being used in other ways here and downstream if desired. The creator is grayscaledrake.