Adds a persistent notes feature (#16923)
* Adds a persistent notes feature

Added a new "Private Notes" option to character setup, and a new "Private Notes" verb that opens them up and allows you to see and edit them in-game in the same way as OOC Notes.

These notes are simply a block of persistent text, purely for your own viewing. The entire purpose of them is for making notes for your own use, such as adding info if you come up with lore on the fly, so that you can easily reference it in the future. You can obviously store anything in there that you'd normally put into a txt file, it's just an easy way to organise info directly tied to your character. This works on a character by character basis.

I've tried to follow the template used for OOC Notes as best I can, because I believe that shadow has updated those to be tg style prefs, so that it'll not need to be redone.

I know that there are a few instances of usr instead of user in here too, I am not sure if they are appropriate or not as I just followed the way that similar settings work. Please do poke me if that needs changing.

* Fixed requested usr's

* Removed to_world

* removed "as mob"