Reflexive Biting Trait
Added a new trait that causes you to bite the hand of anyone who either pats you on the head or boops you on the nose. This deals one point of damage to the hand committing said offense. This also includes a verb in Abilities to toggle this off and on.