adds floor tiles to sheet loader gripper
I had this cool idea for a like, storage box of tiles that you could kinda just slurp up and store but I realized that was really complex and all borgs needed was literally any method of interacting with stacks of floor tiles beyond visual contempt. so this is that.